Webtorrent Implementation and code cleanup. (#8)
* webtorrent implementation

* removed pieces of unneeded code

* minor changes

* destroy client on cancel connection

* used cdn sources and deleted local files

* gitignore added and ide files removed

* lint rules followed

* progress bar not working on sender's side fixed

* Download button not showing fixed.

This was caused because torrent was removed before blob url was generated. Fixed by calling the client.destroy() inside the getBlogUrl() callback function.

* UI fix

* download button not containing any text fixed

* Minor bug fix

when two users were connected and if the user who made the connection request disconnected then the other user was not getting prompted and the screen wasn't being reset to online users page.