Refactored client side code; Added a socket event
It was divided into files, with each file mostly dealing with setting
up listeners for a page and providing methods to show the page. This
would allow us to go back and forth to these pages.

Added an event which would send information to all users who have sent
offers to a user when that user disconnects